feat: passage en composant
