+++ ui/components/SearchForTrainingsAndJobs/components/ResultLists.js
@@ -1,58 +1,32 @@
 import React, { useContext, useState } from "react"
-import { useRouter } from "next/router"
+
 import { ErrorMessage } from "../../../components"
 import { DisplayContext } from "../../../context/DisplayContextProvider"
 import { ScopeContext } from "../../../context/ScopeContext"
 import { SearchResultContext } from "../../../context/SearchResultContextProvider"
-import { ParameterContext } from "../../../context/ParameterContextProvider"
 import { isCfaEntreprise } from "../../../services/cfaEntreprise"
 import { mergeJobs, mergeOpportunities } from "../../../utils/itemListUtils"
-import { currentSearch } from "../../../utils/currentPage.js"
 import { filterLayers } from "../../../utils/mapTools"
 import DuoContainer from "./DuoContainer"
 import ExtendedSearchButton from "./ExtendedSearchButton"
 import NoJobResult from "./NoJobResult"
 import ResultListsCounter from "./ResultListsCounter"
 import purpleFilterIcon from "../../../public/images/icons/purpleFilter.svg"
-import switchOnImage from "../../../public/images/switch-on.svg"
-import switchOffImage from "../../../public/images/switch-off.svg"
-import { refreshLocationMarkers } from "../../../utils/mapTools"
+import DisplayMapButton from "../../../components/DisplayMapButton/displayMapButton"
 
-import { Box, Button, Flex, FormControl, Image, Text } from "@chakra-ui/react"
+import { Box, Button, Flex, Image } from "@chakra-ui/react"
 import { SendPlausibleEvent } from "../../../utils/plausible"
 import { renderJob, renderLbb, renderTraining } from "../services/renderOneResult"
-import pushHistory from "../../../utils/pushHistory"
 
 const ResultLists = (props) => {
-  const router = useRouter()
 
   const scopeContext = useContext(ScopeContext)
-  const { formValues } = useContext(DisplayContext)
-
+  
   let [extendedSearch, hasSearch, isFormVisible] = [false, false, false]
 
   ;({ isFormVisible } = useContext(DisplayContext))
   ;({ extendedSearch, hasSearch } = useContext(SearchResultContext))
-  const { displayMap, setDisplayMap } = useContext(ParameterContext)
   
-  const toggleMapDisplay = () => {
-    const shouldRefreshLocationMarkers = !displayMap
-    setDisplayMap(shouldRefreshLocationMarkers)
-
-    if(shouldRefreshLocationMarkers) {
-      refreshLocationMarkers( { jobs: props.jobs, trainings: props.trainings, scopeContext } )      
-    }
-
-    pushHistory({
-      router,
-      scopeContext,
-      display: "list",
-      searchParameters: formValues,
-      searchTimestamp: currentSearch,
-      displayMap: shouldRefreshLocationMarkers,
-    })
-  }
-
   if (props.isTestMode) {
     ;[extendedSearch, hasSearch, isFormVisible] = [props.stubbedExtendedSearch, props.stubbedHasSearch, props.stubbedIsFormVisible]
   }
@@ -248,17 +222,8 @@ const ResultLists = (props) => {
             <Image width="24px" height="24px" src={purpleFilterIcon} alt="" />
           </Button>
         </Flex>
-        <Flex flex="1 auto" mt={[0,0,2]} alignItems="center" justifyContent="flex-end" display={["none", "none", "flex"]}>
-          <FormControl  flex="0" justifyContent="flex-end" alignItems='center'>
-            <Button mr={[4,4,4,12]} mt={0} display='flex' _hover={{ bg: "none" }} _focus={{ bg: "none" }} background="none" border="none" onClick={toggleMapDisplay}>
-              <Text as="span" fontWeight={400} mr={8} mb='0' fontSize="1rem" >
-                Afficher la carte
-              </Text>
-              {" "}
-              <Image mb="2px" mr="5px" src={displayMap ? switchOnImage : switchOffImage} alt={`Cliquer pour ${displayMap?"masquer":"afficher"} la carte`} />
-            </Button>
-          </FormControl>
-        </Flex>
+        <DisplayMapButton jobs={props.jobs}
+                          trainings={props.trainings} />
         <ResultListsCounter
           scopeContext={scopeContext}
           filterButtonClicked={filterButtonClicked}
